refactor(signer): feature gate bls & bn254 to their own submodules
