+++ dsigner/examples/dsigner_http/main.rs
@@ -7,10 +7,10 @@ use axum::http::StatusCode;
 use axum::response::{IntoResponse, Response};
 use axum::routing::get;
 use axum::{Json, Router, extract::State, routing::post};
-use dcipher_agents::signer::threshold_signer::{
-    AsyncThresholdSigner, ThresholdSigner, lagrange_points_interpolate_at,
+use dcipher_agents::signer::AsynchronousSigner;
+use dcipher_agents::signer::bls::{
+    AsyncThresholdSigner, BN254SignatureOnG1Signer, ThresholdSigner, lagrange_points_interpolate_at,
 };
-use dcipher_agents::signer::{AsynchronousSigner, BN254SignatureOnG1Signer};
 use dcipher_network::transports::libp2p::{Libp2pNode, Libp2pNodeConfig};
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
