+++ randomness-agent/examples/randomness/main.rs
@@ -13,8 +13,8 @@ use dcipher_agents::agents::randomness::fulfiller::RandomnessFulfiller;
 use dcipher_agents::fulfiller::{RequestChannel, Stopper, TickerBasedFulfiller};
 use dcipher_agents::signature_sender::contracts::SignatureSender;
 use dcipher_agents::signature_sender::{SignatureRequest, SignatureSenderFulfillerConfig};
-use dcipher_agents::signer::BN254SignatureOnG1Signer;
-use dcipher_agents::signer::threshold_signer::ThresholdSigner;
+use dcipher_agents::signer::bls::BN254SignatureOnG1Signer;
+use dcipher_agents::signer::bls::ThresholdSigner;
 use dcipher_network::transports::libp2p::{Libp2pNode, Libp2pNodeConfig};
 use randomness_agent::{NotifyTicker, RANDOMNESS_SCHEME_ID, run_agent};
 use std::time::Duration;
