+++ signer/Cargo.toml
@@ -4,13 +4,15 @@ version.workspace = true
 edition.workspace = true
 
 [features]
+bls = []
+bn254 = ["bls", "dep:ark-bn254"]
 rayon = ["dep:rayon"]
 
 [dependencies]
 dcipher-network.workspace = true
 
 # crypto
-ark-bn254 = { workspace = true }
+ark-bn254 = { workspace = true, optional = true }
 ark-ec = { workspace = true }
 ark-ff = { workspace = true }
 ark-std = { workspace = true }
